@@ -44,28 +44,30 @@
     "cleanup:hard": "rm -rf ./dist && rm -rf ./node_modules && rm -rf ./package-lock.json"
   },
   "dependencies": {
-    "vue": "^2.7.10"
+    "vue": "^3.4.15"
   },
   "peerDependencies": {
     "axios": "^1.3.3",
     "lodash": "^4.17.21",
     "luxon": "^3.2.1",
     "require-css": "^0.1.10",
-    "vue": "^2.7.10"
+    "vue": "^3.4.15"
   },
   "devDependencies": {
     "@babel/core": "^7.19.3",
     "@babel/preset-env": "^7.19.4",
-    "@rollup/plugin-alias": "^4.0.3",
-    "@rollup/plugin-babel": "^6.0.0",
-    "@vitejs/plugin-vue2": "^2.0.0",
+    "@rollup/plugin-alias": "^5.1.0",
+    "@rollup/plugin-babel": "^6.0.4",
+    "@vitejs/plugin-vue": "^5.0.3",
     "glob": "^8.0.3",
     "install": "^0.13.0",
     "lodash": "*",
     "npm": "^8.19.2",
     "rollup": "^2.79.1",
-    "rollup-plugin-vue": "^5.1.9",
-    "vite": "^3.1.0",
-    "vue-template-compiler": "^2.7.10"
+    "rollup-plugin-scss": "^4.0.0",
+    "rollup-plugin-vue": "^6.0.0",
+    "sass": "^1.70.0",
+    "scss": "^0.2.4",
+    "vite": "^5.0.11"
   }
 }
